chore: remove deprecated run_analysis_in_process API
- Replace run_analysis_in_process with _run_analysis_async in commands.py
- _run_analysis_async uses callback-based start_analysis() API
- Uses asyncio.Event instead of threading.Event for cleaner async code
- Update tests to mock _run_analysis_async instead of run_analysis_in_process
- Delete deprecated run_analysis_in_process from worker.py
